spider-robot/requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
boto==2.49.0
boto3==1.34.46
botocore==1.34.46
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
cli-helpers==2.3.1
click==8.1.6
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
dnspython==2.6.1
duplicity==2.1.4
evdev==1.7.0
fasteners==0.18
frozenlist==1.4.0
greenlet==3.0.3
httplib2==0.20.4
idna==3.6
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==24.3.1
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
msgpack==1.0.3
multidict==6.0.4
mutagen==1.46.0
mycli==1.27.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.38.2
proton-vpn-gtk-app==4.8.1
proton-vpn-network-manager==0.10.1
psutil==5.9.8
ptyprocess==0.7.0
pyaes==1.6.1
pycairo==1.25.1
pycares==4.4.0
pycryptodomex==3.20.0
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pynvim==0.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-gnupg==0.5.2
python-xlib==0.33
pythondialog==3.5.1
pytz==2024.1
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.3
sentry-sdk==1.39.2
setuptools==68.1.2
six==1.16.0
sqlglot==22.2.0
sqlparse==0.4.4
systemd-python==235
tabulate==0.8.10
terminaltables==3.1.10
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
wcwidth==0.2.5
websockets==10.4
wheel==0.42.0
xdg==5
xkit==0.0.0
yarl==1.8.2
yt-dlp==2024.4.9
zipp==1.0.0
